add auth

+++ src/app/categories/categories.controller.ts
@@ -11,6 +11,8 @@ import { CategoriesService } from './categories.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { ApiTags } from '@nestjs/swagger';
+import { Org, Roles, User } from '@/security/user.decorator';
+import { Role } from '@/security/role.decorator';
 
 @Controller('categories')
 @ApiTags('Service Category')
@@ -18,13 +20,15 @@ export class CategoriesController {
   constructor(private readonly categoriesService: CategoriesService) {}
 
   @Post()
-  create(@Body() createCategoryDto: CreateCategoryDto) {
-    return this.categoriesService.create(createCategoryDto);
+  @Role(Roles.member, Roles.org)
+  create(@Body() createCategoryDto: CreateCategoryDto, @Org() orgId: number) {
+    return this.categoriesService.create(createCategoryDto, orgId);
   }
 
   @Get()
-  findAll() {
-    return this.categoriesService.findAll();
+  @Role(Roles.member, Roles.org)
+  findAll(@Org() orgId: number) {
+    return this.categoriesService.findAll(orgId);
   }
 
   @Get(':id')
@@ -33,15 +37,18 @@ export class CategoriesController {
   }
 
   @Patch(':id')
-  update(
+  async update(
     @Param('id') id: string,
     @Body() updateCategoryDto: UpdateCategoryDto,
+    @Org() orgId: number,
   ) {
+    await this.categoriesService.checkOwnership(+id, orgId);
     return this.categoriesService.update(+id, updateCategoryDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @Org() orgId: number) {
+    await this.categoriesService.checkOwnership(+id, orgId);
     return this.categoriesService.remove(+id);
   }
 }
